+++ src/Files/profile.ps1
@@ -1,49 +1,8 @@
-function prompt {
-    # .Description
-    # This custom version of the PowerShell prompt will present a colorized location value based on the current provider. It will also display the PS prefix in red if the current user is running as administrator.    
-    # .Link
-    # https://go.microsoft.com/fwlink/?LinkID=225750
-    # .ExternalHelp System.Management.Automation.dll-help.xml
-
-    $adminfg = switch ($script:MyOS) {
-        "Windows" {
-            $user = [Security.Principal.WindowsIdentity]::GetCurrent()
-            switch ((New-Object Security.Principal.WindowsPrincipal $user).IsInRole([Security.Principal.WindowsBuiltinRole]::Administrator)) {
-                $true {
-                    "Red"
-                }
-                $false {
-                    "White"
-                }
-            }
-        }
-        default {
-            "White"
-        }
-    }
-
-    switch ((Get-Location).Provider.Name) {
-        "FileSystem"    { $fg = "green"}
-        "Registry"      { $fg = "magenta"}
-        "wsman"         { $fg = "cyan"}
-        "Environment"   { $fg = "yellow"}
-        "Certificate"   { $fg = "darkcyan"}
-        "Function"      { $fg = "gray"}
-        "alias"         { $fg = "darkgray"}
-        "variable"      { $fg = "darkgreen"}
-        default         { $fg = $host.ui.rawui.ForegroundColor}
-    }
-
-    Write-Host ("[{0}@{1}] " -f $global:MyUsername, [System.Net.Dns]::GetHostName()) -NoNewline
-    Write-Host ("[{0}]" -f (Get-Date -Format "HH:mm:ss")) -NoNewline
-    Write-Host " PS " -NoNewline -ForegroundColor $adminfg
-    Write-Host $ExecutionContext.SessionState.Path.CurrentLocation -ForegroundColor $fg
-    Write-Output ("{0} " -f (">" * ($nestedPromptLevel + 1)))
-}
-
 function Update-Profile {
     $Module = Import-Module codaamok -PassThru -ErrorAction "Stop"
-    Copy-Item -Path "$($Module.ModuleBase)\profile.ps1" -Destination $profile.CurrentUserAllHosts -Confirm -ErrorAction "Stop"
+    Copy-Item -Path "$($Module.ModuleBase)\profile.ps1" -Destination $profile.CurrentUserAllHosts -Force -ErrorAction "Stop"
+    $OhMyPoshTheme = Copy-Item -Path "$($Module.ModuleBase)\M365Princess.omp.json" -Destination $HOME -Force -PassThru -ErrorAction "Stop"
+    $profile | Add-Member -MemberType "NoteProperty" -Name "OhMyPoshTheme" -Value $OhMyPoshTheme.FullName
     '. $profile.CurrentUserAllHosts' | clip
     Write-Host "Paste your clipboard"
 }
@@ -54,7 +13,9 @@ function Update-ProfileModule {
         $Available = Find-Module -Name "codaamok" -ErrorAction "Stop"
 
         if ($Installed[0].Version -ne $Available.Version) {
-            Update-Module -Name "codaamok" -ErrorAction "Stop"
+            $Module = Update-Module -Name "codaamok" -Force -PassThru -ErrorAction "Stop"
+            Copy-Item -Path "$($Module.ModuleBase)\profile.ps1" -Destination $profile.CurrentUserAllHosts -Force -ErrorAction "Stop"
+            Copy-Item -Path "$($Module.ModuleBase)\M365Princess.omp.json" -Destination $HOME -Force -ErrorAction "Stop"
         }
     }
 
@@ -100,23 +61,6 @@ function Get-Username {
     }
 }
 
-# Do not use OneDrive\Documents folder for PSModulePath
-
-$env:PSModulePath = ($env:PSModulePath -split ';' | ForEach-Object {
-    if ($_ -notmatch 'OneDrive') { $_ }
-    else { 
-        $_ -replace [Regex]::Escape($_), $(
-            if ($PSVersionTable.PSVersion -ge [System.Version]"7.0") {
-                if ($IsLinux) { "{0}/.local/share/powershell/Modules" -f $home }
-                else { "{0}\Documents\PowerShell\Modules\" -f $home }
-            }
-            else {
-                "{0}\Documents\WindowsPowerShell\Modules\" -f $home
-            }
-        )
-    }
-}) -join ';'
-
 if (-not (Get-Module "codaamok" -ListAvailable)) {
     $answer = Read-Host -Prompt "Profile module not installed, install? (Y)"
     if ($answer -eq "Y" -or $answer -eq "") {
@@ -124,6 +68,8 @@ if (-not (Get-Module "codaamok" -ListAvailable)) {
     }
 }
 else {
+    $Module = Import-Module codaamok -PassThru -ErrorAction "Stop"
+    oh-my-posh prompt init pwsh --config 'C:\Users\AdamCook\M365Princess.omp.json' | Invoke-Expression
     Update-ProfileModule
 }
 
@@ -148,6 +94,13 @@ $script:WorkApps = @(
 
 if (Test-Path $script:machineprofile) {
     . $script:machineprofile
+}
+
+if (Get-Command "oh-my-posh") {
+    oh-my-posh prompt init pwsh --config $profile.OhMyPoshTheme | Invoke-Expression
+}
+else {
+    "oh-my-posh is not installed"
 }
 
 Set-Alias -Name "ctj" -Value "ConvertTo-Json"